Move logic to base class
